+++ src/icemac/ab/calendar/browser/resource.py
@@ -8,6 +8,7 @@ import zope.viewlet.viewlet
 
 lib = fanstatic.Library('calendar', 'resources')
 calendar_css = fanstatic.Resource(lib, 'calendar.css', depends=[base_css])
+print_css = fanstatic.Resource(lib, 'print.css', depends=[calendar_css])
 calendar_js = fanstatic.Resource(
     lib, 'calendar.js', depends=[js.jquery.jquery, js.classy.classy],
     bottom=True)
@@ -18,6 +19,7 @@ class CalendarResources(zope.viewlet.viewlet.ViewletBase):
 
     def update(self):
         calendar_css.need()
+        print_css.need()
         calendar_js.need()
 
     def render(self):