+++ src/icemac/ab/calendar/browser/event.py
@@ -209,8 +209,12 @@ class RecurredEventFormMixIn(object):
         recurring_event = recurring_events[
             self.session['recurring-event-name']]
         date = self.session['recurred-event-date']
-        datetime = recurring_event.datetime.replace(
-            year=date.year, month=date.month, day=date.day)
+        timezone = icemac.addressbook.preferences.utils.get_time_zone()
+        datetime = timezone.normalize(recurring_event.datetime)
+        datetime = datetime.replace(
+            year=date.year, month=date.month, day=date.day, tzinfo=None)
+        # Fix possible DST offset:
+        datetime = timezone.localize(datetime)
         data = icemac.ab.calendar.event.get_event_data_from_recurring_event(
             recurring_event, datetime)
         data['datetime'] = Datetime(